Comparisons/PyForecastDocker/app/requirements.txt
asn1crypto==0.24.0
azure-common==1.1.16
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
cycler==0.10.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
patsy==0.5.1
pycparser==2.19
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
scikit-learn==0.20.1
scipy==1.1.0
six==1.12.0
sklearn==0.0
statsmodels==0.9.0
urllib3==1.24.1
Werkzeug==0.14.1
